+++ backend/src/main/java/com/acervi/backend/controller/AssuntoTopicoController.java
@@ -1,5 +1,21 @@
 package com.acervi.backend.controller;
 
+import java.util.Arrays;
+import java.util.List;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.acervi.backend.model.AssuntoTopico;
+
+@RestController
+@RequestMapping("/api/enum")
 public class AssuntoTopicoController {
     
+    @GetMapping("/650")
+    public List<AssuntoTopico> getAssuntoTopico() {
+        return Arrays.asList(AssuntoTopico.values());
+    }
+
 }
