--- backend/src/main/java/com/acervi/backend/controller/CuttterController.java
@@ -1,5 +1,0 @@
-package com.acervi.backend.controller;
-
-public class CuttterController {
-    
-}
