+++ backend/src/main/java/com/acervi/backend/controller/LivroController.java
@@ -23,6 +23,13 @@ public class LivroController {
 
     @Autowired
     private LivroService livroService;
+
+    // Endpoints:
+    // POST /livros - Criar um livro. check
+    // GET /livros - Listar todos os livros. check
+    // GET /livros/{id} - Buscar um livro pelo ID. check
+    // PUT /livros/{id} - Editar um livro. check
+    // DELETE /livros/{id} - Excluir um livro. check
     
     @Operation(summary = "Criar um novo livro")
     @PostMapping
@@ -31,7 +38,7 @@ public class LivroController {
     }
 
     @Operation(summary = "Buscar um livro pelo ID")
-    @PostMapping("/{id}")
+    @GetMapping("/{id}")
     public Livro buscarLivro(@PathVariable Long id) {
         return livroService.buscarLivro(id);
     }
